feat/sword Change file management
