+++ src/main/java/fr/loanspac/magicstuff/skill/Skill.java
@@ -1,0 +1,16 @@
+package fr.loanspac.magicstuff.skill;
+
+import fr.loanspac.magicstuff.utils.CooldownManager;
+import lombok.Getter;
+import org.bukkit.entity.Player;
+
+@Getter
+public abstract class Skill {
+    private final String name;
+    private final CooldownManager cooldown = new CooldownManager();
+    public abstract void executor(Player player);
+
+    protected Skill(String name) {
+        this.name = name;
+    }
+}
